feat: handle drop index
