+++ internal/pkg/incr/applyops_test.go
